add 965. Univalued Binary Tree

+++ Leetcodecpp/MyStack.h
@@ -7,7 +7,7 @@ public:
   MyStack() {}
 
   void push(int x) {
-    int n = q.size();
+    size_t n = q.size();
     q.push(x);
     for (int i = 0; i < n; i++) {
       q.push(q.front());
