+++ Leetcodecpp/NumArray.h
@@ -3,7 +3,7 @@
 class NumArray {
 public:
   NumArray(std::vector<int>& nums) {
-    int len = nums.size();
+    size_t len = nums.size();
     sum = std::vector<int>(len + 1);
     for (int i = 0; i < len; i++) {
       sum[i + 1] = sum[i] + nums[i];
