+++ Leetcodecpp/RandomizedSet.cpp
@@ -4,7 +4,7 @@ bool RandomizedSet::insert(int val) {
   if (indices.count(val)) {
     return false;
   }
-  size_t index = nums.size();
+  int index = (int)nums.size();
   nums.emplace_back(val);
   indices[val] = index;
 	return true;
@@ -14,7 +14,7 @@ bool RandomizedSet::remove(int val) {
   if (!indices.count(val)) {
     return false;
   }
-  size_t index = indices[val];
+  int index = (int)indices[val];
   int last = nums.back();
   nums[index] = last;
   indices[last] = index;
@@ -26,6 +26,6 @@ bool RandomizedSet::remove(int val) {
 int RandomizedSet::getRandom() { 
   std::random_device r;
   std::mt19937 el(r());
-  std::uniform_int_distribution<int> uniform_dist(0, nums.size() - 1);
+  std::uniform_int_distribution<int> uniform_dist(0, int(nums.size()) - 1);
   return nums[uniform_dist(el)];
 }
